% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{angdist}
\alias{angdist}
\title{Calculates angular distance theta between points on a sphere specified by latitude and longitude}
\usage{
angdist(lat1, lon1, lat2, lon2)
}
\arguments{
\item{lat1}{latitude of point 1 (degrees)}

\item{lon1}{longitude of point 1 (degrees)}

\item{lat2}{latitude of point 2 (degrees)}

\item{lon2}{longitude of point 2 (degrees)}
}
\description{
Calculates angular distance theta between points on a sphere specified by latitude and longitude
}
