% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{atm_pressure}
\alias{atm_pressure}
\title{Compute atmospheric pressure at site}
\usage{
atm_pressure(alt, lon = NULL, lat = NULL, model = "stone2000")
}
\arguments{
\item{alt}{Altitude (m)}

\item{lon}{Longitude (degrees) optional (only for "era40" model)}

\item{lat}{Latitude (degrees) optional (only for "era40" model)}

\item{model}{Model used for computation, one of "stone2000" or "era40"}
}
\value{
Atmospheric pressure in hPa
}
\description{
Compute atmospheric pressure at site
}
\details{
Choice of models :

(1) "stone2000" (default) : equation 1 of Stone (2000) JGR paper (only function of site altitude)
Stone (2000)  https://doi.org/10.1029/2000JB900181

(2) "era40" : Looks up mean sea level pressure and mean 1000 mb temp from ERA-40 reanalysis
and calculates site atmospheric pressures using these as inputs to the standard atmosphere equation
(function of site latitude, longitude and altitude).
Uppala et al (2005)  https://doi.org/10.1256/qj.04.176
}
\examples{
z = seq(0,5000,by=100)
P1=atm_pressure(alt=z,model="stone2000")
P2=atm_pressure(alt=z,lat=-60,lon=0,model="era40")
plot(P1,z,type="l",xlab="Pressure (hPa)",ylab="Altitude (m)")
lines(P2,z,lty=2)
}
