% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_mixing.R
\name{conc_soil_mixing}
\alias{conc_soil_mixing}
\title{Nuclide concentration in a vertically mixed soil}
\usage{
conc_soil_mixing(h, E, rhos, rhob, p, L, S)
}
\arguments{
\item{h}{soil thickness (cm)}

\item{E}{erosion rate (m/Ma)}

\item{rhos}{soil density (g/cm3)}

\item{rhob}{bedrock density (g/cm3)}

\item{p}{production and decay parameters (4 elements vector)
p[1] -> unscaled spallation production rate (at/g/a)
p[2] -> unscaled stopped muons production rate (at/g/a)
p[3] -> unscaled fast muons production rate (at/g/a)
p[4] -> decay constant (1/a)}

\item{L}{attenuation lengths (3 elements vector)
L[1] -> neutrons
L[2] -> stopped muons
L[3] -> fast muons}

\item{S}{scaling factors (2 elements vector)
S[1] -> scaling factor for spallation
S[2] -> scaling factor for muons}
}
\value{
Concentration in at/g
}
\description{
Compute nuclide concentration in a vertically-mixed steady-state soil, using the approach of Foster et al. (2015)
}
