% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_mixing.R
\name{depth_averaged_prod}
\alias{depth_averaged_prod}
\title{Average nuclide production rate in soil}
\usage{
depth_averaged_prod(h, p, L, S)
}
\arguments{
\item{h}{soil thickness (in g/cm2)}

\item{p}{production and decay parameters (4 elements vector)
p[1] -> unscaled spallation production rate (at/g/a)
p[2] -> unscaled stopped muons production rate (at/g/a)
p[3] -> unscaled fast muons production rate (at/g/a)
p[4] -> decay constant (1/a)}

\item{L}{attenuation lengths (3 elements vector)
L[1] -> neutrons
L[2] -> stopped muons
L[3] -> fast muons}

\item{S}{scaling factors (2 elements vector)
S[1] -> scaling factor for spallation
S[2] -> scaling factor for muons}
}
\value{

}
\description{
Compute average nuclide production rates for vertically mixed soil column
}
