% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth_profile.R
\name{depth_profile_mc}
\alias{depth_profile_mc}
\title{Monte Carlo exploration for a depth profile}
\usage{
depth_profile_mc(C, C_e, depth, prm, L, S, age, ero, inh, rho, n1 = 0, n2 = 0)
}
\arguments{
\item{C}{concentrations vector (at/g)}

\item{C_e}{uncertainties on the concentrations vector (at/g)}

\item{prm}{production and decay parameters (4 elements vector)
p[1] -> unscaled spallation production rate (at/g/a)
p[2] -> unscaled stopped muons production rate (at/g/a)
p[3] -> unscaled fast muons production rate (at/g/a)
p[4] -> decay constant (1/a)}

\item{L}{attenuation lengths (3 elements vector)
L[1] -> neutrons
L[2] -> stopped muons
L[3] -> fast muons}

\item{S}{scaling factors (2 elements vector)
S[1] -> scaling factor for spallation
S[2] -> scaling factor for muons}

\item{age}{age range to explore (a). 2 elements vector. If 1 element this parameter is fixed}

\item{ero}{erosion rate range to explore (m/Ma). 2 elements vector. If 1 element this parameter is fixed}

\item{inh}{inherited concentration range to explore (at/g). 2 elements vector. If 1 element this parameter is fixed}

\item{rho}{density range to explore (g/cm3). 2 elements vector. If 1 element this parameter is fixed}

\item{n1}{number of runs with random}

\item{n2}{}

\item{z}{depth vector (g/cm2)}
}
\description{
Eulerian approach
}
