% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{fun_solv_ss_eros_eul}
\alias{fun_solv_ss_eros_eul}
\title{Function to be solved by solv_ss_eros_eul when method="single"}
\usage{
fun_solv_ss_eros_eul(ero, Cobs, p, S, L)
}
\arguments{
\item{L}{}
}
\description{
Function to be solved by solv_ss_eros_eul when method="single"
}
