% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_vdm}
\alias{get_vdm}
\title{Get Virtual Dipole Moment time series}
\usage{
get_vdm(time, model)
}
\arguments{
\item{time}{time (a)}

\item{model}{model to be used (one of musch, glopis or lsd)}
}
\description{
Original data from Crep calculator (See Martin et al., 2017)
}
\examples{
get_vdm
}
