% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Lambda}
\alias{Lambda}
\title{A set of attenuation lengths for cosmogenic nuclides production pathways}
\format{
A vector with 3 elements
\describe{
  \item{Lspal}{Spallogenic neutrons attenuation length (g/cm2)}
  \item{Lstop}{Stopping muons attenuation length (g/cm2)}
  \item{Lfast}{Fast muons attenuation length (g/cm2)}
}
}
\usage{
Lambda
}
\description{
A set of attenuation lengths for cosmogenic nuclides production pathways
}
\keyword{datasets}
