% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{likelihood1}
\alias{likelihood1}
\title{Likelihood function : P(obs|params) for solv_ss_eros_eul when method="MCMC"}
\usage{
likelihood1(ero, p, S, L, Cobs, Cobs_e)
}
\description{
Likelihood function : P(obs|params) for solv_ss_eros_eul when method="MCMC"
}
