% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_lsd.R
\name{lsdrc}
\alias{lsdrc}
\title{Compute cutoff rigidities time series for the calculation of scaling factors according to Lifton et al. (2014)
'LSD' scaling scheme}
\usage{
lsdrc(lat, lon, time)
}
\arguments{
\item{lat}{site latitude (deg, -90->90 southern hemisphere negative)}

\item{lon}{site longitude (deg, -180->180 western hemisphere negative)}

\item{time}{time before present (a, with respect to 2010)}
}
\description{
Adapted from the matlab code in the supplementary material of Lifton et al. (2014)
}
