% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{posterior1}
\alias{posterior1}
\title{Un-normalized posterior : numerator of Bayes formula : P(obs|params)*P(params) for solv_ss_eros_eul when method="MCMC"}
\usage{
posterior1(ero, p, S, L, Cobs, Cobs_e, binf, bsup)
}
\description{
Un-normalized posterior : numerator of Bayes formula : P(obs|params)*P(params) for solv_ss_eros_eul when method="MCMC"
}
\keyword{internal}
