% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{prior1}
\alias{prior1}
\title{Prior distribution  P(params) for solv_ss_eros_eul when method="MCMC"}
\usage{
prior1(ero, binf, bsup)
}
\description{
Prior distribution  P(params) for solv_ss_eros_eul when method="MCMC"
}
