% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prm}
\alias{prm}
\title{A set of reference parameters for SLHL production rates and decay constant}
\format{
A matrix with nuclides in columns and production rates or decay constant in rows
\describe{
  \item{Pspal}{SLHL spallation production rate for scaling scheme st (at/g/a)}
  \item{Pstop}{SL stopping muons production rate (at/g/a)}
  \item{Pfast}{SL fast muons production rate (at/g/a)}
  \item{lambda}{Decay constant (1/a)}
}
}
\usage{
prm
}
\description{
10Be, 26Al and 14C spallation production rates from Borchers et al. (2016).
10Be and 26Al muons production rates are from Braucher et al. (2011).
14C muons production rate is from Lupker et al. (2015).
}
\details{
Borchers, B., Marrero, S., Balco, G., Caffee, M., Goehring, B., Lifton, N., Nishiizumi, K., Phillips, F., Schaefer, J., & Stone, J. (2016). Geological calibration of spallation production rates in the CRONUS-Earth project. Quaternary Geochronology, 31, 188–198. https://doi.org/10.1016/j.quageo.2015.01.009

Braucher, R., Merchel, S., Borgomano, J., & Bourlès, D. L. (2011). Production of cosmogenic radionuclides at great depth: A multi element approach. Earth and Planetary Science Letters, 309(1–2), 1–9. http://dx.doi.org/10.1016/j.epsl.2011.06.036

Lupker, M., Hippe, K., Wacker, L., Kober, F., Maden, C., Braucher, R., Bourlès, D., Romani, J. R. V., & Wieler, R. (2015). Depth-dependence of the production rate of in situ14C in quartz from the Leymon High core, Spain. Quaternary Geochronology, 28, 80–87. https://doi.org/10.1016/j.quageo.2015.04.004
}
\keyword{datasets}
