% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{proposal1}
\alias{proposal1}
\title{Proposal function for solv_ss_eros_eul when method="MCMC"}
\usage{
proposal1(ero, binf, bsup)
}
\description{
Proposal function for solv_ss_eros_eul when method="MCMC"
}
