% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rc_ndp}
\alias{rc_ndp}
\title{calculate cutoff rigidity from non-dipolar gridded model}
\usage{
rc_ndp(time, lat, lon)
}
\arguments{
\item{time}{time (a)}

\item{lat}{Latitude (deg)}

\item{lon}{Longitude (deg)}
}
\value{

}
\description{
calculate cutoff rigidity from non-dipolar gridded model
}
\examples{
vdm2rc
}
