% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{run_mcmc1}
\alias{run_mcmc1}
\title{RUN MCMC function for solv_ss_eros_eul when method="MCMC"}
\usage{
run_mcmc1(p, S, L, Cobs, Cobs_e, binf, bsup, n)
}
\description{
RUN MCMC function for solv_ss_eros_eul when method="MCMC"
}
