% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sato_neutrons_low.R
\name{sato_neutrons_low}
\alias{sato_neutrons_low}
\title{Compute low energy neutron flux using the analytical formulation of Sato et al. (2008)}
\usage{
sato_neutrons_low(h, Rc, s, w)
}
\arguments{
\item{h}{atmospheric pressure (hPa)}

\item{Rc}{cutoff rigidity (GV)}

\item{s}{solar modulation potential}

\item{w}{gravimetric fractional water content (non-dimensional)}
}
\description{
Compute low energy neutron flux using the analytical formulation of Sato et al. (2008)
}
\details{
Adaption of matlab code from Lifton et al. (2014)

Lifton, N., Sato, T., & Dunai, T. J. (\strong{2014}). Scaling in situ cosmogenic nuclide production rates using analytical approximations to atmospheric cosmic-ray fluxes.
\emph{Earth and Planetary Science Letters}, 386, 149–160.
https://doi.org/10.1016/j.epsl.2013.10.052
}
\examples{

}
