% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_st_lm.R
\name{scaling_lm}
\alias{scaling_lm}
\title{Compute scaling parameter for the lm scheme (Lal/Stone time dependent)}
\usage{
scaling_lm(h, Rc)
}
\arguments{
\item{h}{atmospheric pressure (hPa)}

\item{Rc}{cutoff rigidity (GV)}
}
\description{
Adapted from G. Balco matlab code (http://hess.ess.washington.edu/math)
}
\examples{
scaling_lm
}
