% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_lsd.R
\name{scaling_lsd}
\alias{scaling_lsd}
\title{scaling_lsd}
\usage{
scaling_lsd(h, Rc, SPhi, w)
}
\arguments{
\item{h}{atmospheric pressure (hPa)}

\item{Rc}{cutoff rigidity (GV)}

\item{SPhi}{solar modulation potential (Phi, see source paper)}

\item{w}{fractional water content of ground (nondimensional)}
}
\description{
Implements the Lifton-Sato-Dunai scaling scheme from Lifton et al. (2014)
}
\details{
Compute the scaling paramters for both the Sf (flux based) and Sa (flux ad cross-section based, nuclide dependent) schemes.

Lifton, N., Sato, T., & Dunai, T. J. (2014).
Scaling in situ cosmogenic nuclide production rates using analytical approximations to atmospheric cosmic-ray fluxes.
Earth and Planetary Science Letters, 386, 149–160.
https://doi.org/10.1016/j.epsl.2013.10.052

Initial Matlab code from Lifton et al. (2014)
}
