% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_st_lm.R
\name{scaling_st}
\alias{scaling_st}
\title{Compute scaling parameters according to Stone (2000) (st)}
\usage{
scaling_st(P, lat)
}
\arguments{
\item{P}{pressure (hPa)}

\item{lat}{Latitude (deg)}
}
\description{
This function compute spallation and muons scaling factors according to Equations 2 and 3 from Stone (2000)
}
\examples{
scaling_st
}
