% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_mixing.R
\name{soil_mixing_depth}
\alias{soil_mixing_depth}
\title{Steady state soil mixing depth}
\usage{
soil_mixing_depth(C1, C1_e, C2, C2_e, p1, p2, Lambda, S, rhos, rhob)
}
\arguments{
\item{C1}{First nuclide concentration (at/g)}

\item{C1_e}{First nuclide concentration uncertainty (at/g)}

\item{C2}{Second nuclide concentration (at/g)}

\item{C2_e}{Second nuclide concentration uncertainty (at/g)}

\item{p1}{production and decay parameters for first nuclide (4 elements vector)
\itemize{
 \item p1[1] unscaled spallation production rate (at/g/a)
 \item p1[2] unscaled stopped muons production rate (at/g/a)
 \item p1[3] unscaled fast muons production rate (at/g/a)
 \item p1[4] decay constant (1/a)
}}

\item{p2}{production and decay parameters for second nuclide (4 elements vector)
\itemize{
 \item p2[1] unscaled spallation production rate (at/g/a)
 \item p2[2] unscaled stopped muons production rate (at/g/a)
 \item p2[3] unscaled fast muons production rate (at/g/a)
 \item p2[4] decay constant (1/a)
}}

\item{S}{scaling factors (2 elements vector)
\itemize{
\item S[1]  scaling factor for spallation
\item S[2]  scaling factor for muons
}}

\item{L}{Attenuation lengths (3 elements vector in g/cm2)
\itemize{
\item L[1]  neutrons
\item L[2]  stopped muons
\item L[3]  fast muons
}}
}
\value{
A 1 row data frame with the following columns
\itemize{
\item Soil depth (m)
\item Uncertainty in soil depth (m)
\item Erosion rate (m/Ma)
\item Uncertainty on erosion rate (m/Ma)
\item Covariance
}
}
\description{
Compute steady-state soil mixing depth and erosion rate using two nuclides cocnentrations
}
