% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaling_lsd.R
\name{solar_modulation}
\alias{solar_modulation}
\title{Compute solar modulation factor for the LSD scheme}
\usage{
solar_modulation(t)
}
\arguments{
\item{t}{time (a) same as for Rc calculations}
}
\description{
Adapted from the matlab code in the supplementary material of Lifton et al. (2014)
}
