% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{solv_conc_eul}
\alias{solv_conc_eul}
\title{Prediction of concentration evolution  an Eulerian grid}
\usage{
solv_conc_eul(z, ero, t, C0, p, S, L, in_ero = NULL)
}
\arguments{
\item{z}{depth coordinate of the profile (g/cm2)}

\item{ero}{erosion rate (g/cm2/a)}

\item{t}{time interval (a)}

\item{C0}{inherited concentration constant with depth (at/g)}

\item{p}{production and decay parameters (4 elements vector)
p[1] -> unscaled spallation production rate (at/g/a)
p[2] -> unscaled stopped muons production rate (at/g/a)
p[3] -> unscaled fast muons production rate (at/g/a)
p[4] -> decay constant (1/a)}

\item{S}{scaling factors (2 elements vector)
S[1] -> scaling factor for spallation
S[2] -> scaling factor for muons}

\item{L}{attenuation lengths (3 elements vector)
L[1] -> neutrons
L[2] -> stopped muons
L[3] -> fast muons}

\item{in_ero}{initial erosion rate yielding a steady state inherited concentration (g/cm2/a), C0 not used in this case}
}
\description{
Exponential attenuation models for neutrons and muons
}
\details{
If `ero` contains several values, the corresponding evolution in time of concentration is computed, `t` should be of same length. The first value of `ero` is used to compute a steady state initial concentration, then the evolution of concentration is computed by ero[i] is the constant erosion rate over the time interval t[i-1] to t[i].
}
