% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{solv_conc_lag}
\alias{solv_conc_lag}
\title{Compute concentration evolution along a time-depth history}
\usage{
solv_conc_lag(t, z, C0, Psp0, Pmu0, lambda, S, final = FALSE)
}
\arguments{
\item{t}{time vector (a)}

\item{z}{depth vector (g/cm2), same length as t}

\item{C0}{initial concentration (at/g)}

\item{Psp0}{SLHL spallation production profile (at/g/a) at depth corresponding to z}

\item{Pmu0}{muonic production profil at depth corresponding to z}

\item{lambda}{radioactive decay constant (1/a)}

\item{S}{scaling parameters for PsP0 and Pmu0 (2 columns), at times corresponding to t (same number of rows)}

\item{final}{if TRUE only compute the final concentration (default=FALSE)}
}
\description{
Lagrangian formulation
}
