% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvers.R
\name{solv_ss_eros_eul}
\alias{solv_ss_eros_eul}
\title{Compute steady state denudation using Eulerian description for the evolution of concentration}
\usage{
solv_ss_eros_eul(Cobs, p, S, L, Cobs_e = 0, method = "single", n = 10000)
}
\arguments{
\item{Cobs}{Uncertainty on measured concentration (at/g), optional but required ffor method MC and MCMC}

\item{p}{production and decay parameters (4 elements vector)
p[1] -> unscaled spallation production rate (at/g/a)
p[2] -> unscaled stopped muons production rate (at/g/a)
p[3] -> unscaled fast muons production rate (at/g/a)
p[4] -> decay constant (1/a)}

\item{S}{scaling factors (2 elements vector)
S[1] -> scaling factor for spallation
S[2] -> scaling factor for muons}

\item{L}{attenuation length (3 elements vector)
L[1] -> neutrons
L[2] -> stopped muons
L[3] -> fast muons}

\item{method}{One of single (default), MC, MCMC}
}
\description{
Eulerian formulation
}
