% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_nuclides_plots.R
\name{tnp_burial}
\alias{tnp_burial}
\title{Compute burial age and denudation rate curves array for two-nuclides plots}
\usage{
tnp_burial(A, E, p1, p2, L, S, rho, n = 100)
}
\arguments{
\item{A}{burial age vector (a)}

\item{E}{erosion rate vector (m/Ma)}

\item{p1}{production and decay parameters for the first nuclide (4 elements vector)
\itemize{
 \item p1[1] unscaled spallation production rate (at/g/a)
 \item p1[2] unscaled stopped muons production rate (at/g/a)
 \item p1[3] unscaled fast muons production rate (at/g/a)
 \item p1[4] decay constant (1/a)
}}

\item{p2}{production and decay parameters for the second nuclide (4 elements vector, same as p1)}

\item{L}{Attenuation length (3 elements vector in g/cm2)
\itemize{
\item L[1]  neutrons
\item L[2]  stopped muons
\item L[3]  fast muons
}}

\item{S}{scaling factors (2 elements vector)
\itemize{
\item S[1]  scaling factor for spallation
\item S[2]  scaling factor for muons
}}

\item{rho}{material density (g/cm3)}

\item{n}{number of along-curve evaluation points (optional default 100)}
}
\value{
a list with two dataframes containing the concentrations for the two nuclides as a function of burial age and denudation
}
\description{
Compute burial age and denudation rate curves array for two-nuclides plots
}
\examples{
data("prm") # production and decay data
p = prm
data("Lambda") # attenuation length data
L = Lambda
altitude = 1000 # site elevation in m
latitude = 20 # site latitude in degrees
P = atm_pressure(alt=altitude,model="stone2000") # atmospheric pressure at site
S = scaling_st(P,latitude) # Stone 2000 scaling parameters
rho = 2.65 # bedrock density (g/cm3)
N1 = "Be10" # longer half-life
N2 = "Al26" # shorter half-life
res = tnp_curves(prm[,N1],prm[,N2],Lambda,S,rho)
plot(NA,xlim=c(0.75,10),ylim=c(1,7),log="x",
    xlab=paste(N1,"(x10^6 at/g)"),ylab=paste(N2,"/",N1))
lines(res[[1]]$C1/1e6,res[[1]]$C2/res[[1]]$C1,lty=2,lwd=2,col="khaki4") # constant exposure
lines(res[[2]]$C1/1e6,res[[2]]$C2/res[[2]]$C1,lty=1,lwd=2,col="khaki4") # steady-state erosion
A = c(1,2,3,4)*1e6 # increments in age (a)
E = c(0.1,0.2,0.5,1,2,5,10) # increments in denudation rate (m/Ma)
res = tnp_burial(A,E,prm[,N1],prm[,N2],L,S,rho,n=100) # compute array
lines(res[[1]]$C1/1e6,res[[1]]$C2/res[[1]]$C1,col="grey") # constant age
lines(res[[2]]$C1/1e6,res[[2]]$C2/res[[2]]$C1,col="black") # constant denudation
}
