% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_nuclides_plots.R
\name{tnp_curves}
\alias{tnp_curves}
\title{Compute steady state denudation and constant exposure curves for two-nuclides plots}
\usage{
tnp_curves(
  P1,
  P2,
  Lambda,
  S,
  rho,
  dlim = c(0.1, 1000),
  alim = c(100, 1e+07),
  n = 100
)
}
\arguments{
\item{P1}{production and decay parameters for the first nuclide (4 elements vector)
p[1] -> unscaled spallation production rate (at/g/a)
p[2] -> unscaled stopped muons production rate (at/g/a)
p[3] -> unscaled fast muons production rate (at/g/a)
p[4] -> decay constant (1/a)}

\item{P2}{production and decay parameters for the first nuclide (4 elements vector)}

\item{S}{scaling factors (2 elements vector)
S[1] -> scaling factor for spallation
S[2] -> scaling factor for muons}

\item{rho}{density (g/cm3)}

\item{dlim}{range of denudation rates (m/Ma, optional, default 0.1-1000 m/Ma), two elements vector}

\item{alim}{range of exposure ages (a, optional, default 100-10e6 a), two elements vector}

\item{n}{number of evaluation points (optional default 100)}

\item{L}{Attenuation length (3 elements vector)
L[1] -> neutrons
L[2] -> stopped muons
L[3] -> fast muons}
}
\value{
a list with two dataframes containing the concentrations for the two nuclides as a function of denudation rate or exposure age
}
\description{
Compute steady state denudation and constant exposure curves for two-nuclides plots
}
