% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_nuclides_plots.R
\name{tnp_ellipse}
\alias{tnp_ellipse}
\title{Compute uncertainty ellipses for two-nuclides plot}
\usage{
tnp_ellipse(Ca, Ca_e, Cb, Cb_e, confidence = 0.98)
}
\arguments{
\item{Ca}{Concentration of first nuclide (usually 10Be) (at)}

\item{Ca_e}{1-sigma uncertainty on the concentration of first nuclide  (at)}

\item{Cb}{Concentration of second nuclide (usually 26Al) (at)}

\item{Cb_e}{1-sigma uncertainty on the concentration of second nuclide  (at)}

\item{confidence}{confidence level (default 0.98)}
}
\description{
Two nuclides A and B plotted as B/A on the Y-axis and A on the X-axis
Return a list with given confidence level ellipses coordinates for each pair of concentrations
}
