% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_mixing.R
\name{tnp_soil_mixing}
\alias{tnp_soil_mixing}
\title{Compute constant depth and denudation rate curves array for two-nuclides plots}
\usage{
tnp_soil_mixing(h, E, rhos, rhob, p1, p2, L, S, n = 100)
}
\arguments{
\item{h}{soil thickness vector (cm)}

\item{E}{erosion rate vector (m/Ma)}

\item{rhos}{soil density (g/cm3)}

\item{rhob}{bedrock density (g/cm3)}

\item{p1}{production and decay parameters for the first nuclide (4 elements vector)
\itemize{
 \item p1[1] unscaled spallation production rate (at/g/a)
 \item p1[2] unscaled stopped muons production rate (at/g/a)
 \item p1[3] unscaled fast muons production rate (at/g/a)
 \item p1[4] decay constant (1/a)
}}

\item{p2}{production and decay parameters for the second nuclide (4 elements vector, same as p1)}

\item{L}{Attenuation length (3 elements vector in g/cm2)
\itemize{
\item L[1]  neutrons
\item L[2]  stopped muons
\item L[3]  fast muons
}}

\item{S}{scaling factors (2 elements vector)
\itemize{
\item S[1]  scaling factor for spallation
\item S[2]  scaling factor for muons
}}

\item{n}{number of along-curve evaluation points (optional default 100)}
}
\value{
a list with two dataframes containing the concentrations for the two nuclides as a function of soil depth and denudation
}
\description{
Compute constant depth and denudation rate curves array for two-nuclides plots
}
\examples{
data("prm") # production and decay data
p = prm
data("Lambda") # attenuation length data
L = Lambda
altitude = 1000 # site elevation in m
latitude = 20 # site latitude in degrees
P = atm_pressure(alt=altitude,model="stone2000") # atmospheric pressure at site
S = scaling_st(P,latitude) # Stone 2000 scaling parameters
rhob = 2.65 # bedrock density (g/cm3)
rhos = rhob/2 # soil density  (g/cm3)
N1 = "Be10" # longer half-life
N2 = "Al26" # shorter half-life
res = tnp_curves(prm[,N1],prm[,N2],Lambda,S,rhob)
plot(NA,xlim=c(0.75,5),ylim=c(3,7),log="x",
    xlab=paste(N1,"(x10^6 at/g)"),ylab=paste(N2,"/",N1))
lines(res[[1]]$C1/1e6,res[[1]]$C2/res[[1]]$C1,lty=2,lwd=2,col="khaki4") # constant exposure
lines(res[[2]]$C1/1e6,res[[2]]$C2/res[[2]]$C1,lty=1,lwd=2,col="khaki4") # steady-state erosion
h = seq(100,1000,by = 100) # increments in soil depth (cm)
E = c(0.1,0.2,0.5,1,2,5,10) # increments in denudation rate (m/Ma)
res = tnp_soil_mixing(h,E,rhos,rhob,prm[,N1],prm[,N2],L,S,n=100) # compute array
lines(res[[1]]$C1/1e6,res[[1]]$C2/res[[1]]$C1,col="grey") # constant depth
lines(res[[2]]$C1/1e6,res[[2]]$C2/res[[2]]$C1,col="black") # constant denudation
}
