% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vdm2rc}
\alias{vdm2rc}
\title{calculate cutoff rigidity from Virtual Dipole Moment time series}
\usage{
vdm2rc(vdm, lat, model = "elsasser54")
}
\arguments{
\item{vdm}{Virtual Dipole Moment (A.m^2)}

\item{lat}{Latitude (deg)}

\item{model}{Model used to compute Rc from the virtual dipole moment value (one of "elsasser54" or "lifton14")}
}
\value{

}
\description{
calculate cutoff rigidity from Virtual Dipole Moment time series
}
\examples{
vdm2rc
}
